small changes to make tank not get stuck as easily. adding logic to creating shots/bullets to the world

+++ src/TankGame/Shot.java
@@ -8,19 +8,19 @@ public class Shot extends CollidableObject{
     private static final String SHOT_IMAGE = "resources" + File.separator + "Shell_basic_strip60.png";
     private static final int BASE_SPEED = 5;
     private static final int BASE_DAMAGE = 10;
-    private float vector;
+    private int vector;
     private float velocity;
     private int damage;
     
-    public Shot(Point position, float vector) throws IOException{
+    public Shot(Point position, int vector) throws IOException{
         this(position, vector, BASE_SPEED);
     }
     
-    public Shot(Point position, float vector, float velocity) throws IOException{
+    public Shot(Point position, int vector, float velocity) throws IOException{
         this(position, vector, velocity, BASE_DAMAGE);
     }
     
-    public Shot(Point position, float vector, float velocity, int damage) throws IOException{
+    public Shot(Point position, int vector, float velocity, int damage) throws IOException{
         super(position.x, position.y, SHOT_IMAGE);
         this.vector = vector;
         this.velocity = velocity;
@@ -31,7 +31,7 @@ public class Shot extends CollidableObject{
         return vector;
     }
     
-    public void setVector(float newVector) {
+    public void setVector(int newVector) {
         vector = newVector;
     }
